+++ src/command/ECommand.ts
@@ -2,6 +2,7 @@ export const enum ECommand {
     CmdFindClassRef = 'ahk.nekoHelp.CmdFindClassRef',
     CmdFindMethodRef = 'ahk.nekoHelp.CmdFindMethodRef',
     CmdFindFuncRef = 'ahk.nekoHelp.CmdFindFuncRef',
+    CmdFindLabelRef = 'ahk.nekoHelp.CmdFindLabelRef',
     CmdGotoFuncDef = 'ahk.nekoHelp.CmdGotoFuncDef',
     CompletionMsgBox = 'ahk.nekoHelp.CompletionMsgBox', // TODO
     ListAllFunc = 'ahk.nekoHelp.ListAllFunc',