+++ src/provider/CodeLens/ECodeLensStr.ts
@@ -1,0 +1,3 @@
+export const enum ECodeLensStr {
+    tooltip = 'by neko-help dev tools',
+}